+++ test/client-add-duplicate-trackers.js
