+++ test/client-remove.js
@@ -1,0 +1,24 @@
+var common = require('./common')
+var test = require('tape')
+var WebTorrent = require('../')
+
+test('client.remove: remove by Torrent object', function (t) {
+  t.plan(5)
+
+  var client = new WebTorrent({ dht: false, tracker: false })
+
+  client.on('error', function (err) { t.fail(err) })
+  client.on('warning', function (err) { t.fail(err) })
+
+  var torrent = client.add(common.leaves.parsedTorrent.infoHash)
+  t.equal(client.torrents.length, 1)
+
+  torrent.on('infoHash', function () {
+    t.equal(torrent.infoHash, common.leaves.parsedTorrent.infoHash)
+
+    client.remove(torrent, function (err) { t.error(err, 'torrent destroyed') })
+    t.equal(client.torrents.length, 0)
+
+    client.destroy(function (err) { t.error(err, 'client destroyed') })
+  })
+})
